+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Breaking changes
+
+- move pin connecting to timer channels after `Pwm` initialization [#517]
+
 ### Changed
 
 - `timer.rs` refactoring [#553]
@@ -21,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `rtic2` feature, renamed `rtic` to `rtic1`
 - `embedded_io::Read` trait implemented for `serial`
+
+[#517]: https://github.com/stm32-rs/stm32f1xx-hal/pull/517
 
 ## [v0.11.0] - 2025-09-09